arquitectura_1/requirements.txt
litestar
uvicorn
jinja2
base62
grpcio
grpcio-tools